pyproject.toml
[project]
name = "whisper-transcription"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai-whisper>=20250625",
    "setuptools-rust>=1.12.0",
]
